requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
Pillow